Refactored do all CRUD is in one module and updated commenting
