+++ public/views/index.html
@@ -83,7 +83,6 @@
   <!-- Chart.js -->
   <script src="https://cdn.jsdelivr.net/npm/chart.js@2.8.0"></script>
   <!-- Custom js -->
-  <!-- <script>src="../assets/js/index.js"</script> -->
   <script src="../dist/bundle.js"></script>
   <!-- Register the service worker -->
   <script>
